+++ NEWS.md
@@ -1,3 +1,8 @@
+## hutilscpp 0.9.1
+
+* Bug fixes:
+  - 
+
 ## hutilscpp 0.9.0
 
 * New functions: